% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linreg.R
\name{compute_s2g}
\alias{compute_s2g}
\title{Compute Genetic Variance}
\usage{
compute_s2g(RL, beta)
}
\arguments{
\item{RL}{Lower Cholesky factor of the p x p LD matrix for the population.}

\item{beta}{Genotype effects.}
}
\value{
Genetic variance (s2g).
}
\description{
Compute genetic variance given betas and LD Cholesky factor.
}
\details{
s2g := beta' V beta = beta' L L' beta
}
\examples{
R <- matrix(c(1, 1, 1, 1), nrow = 2, ncol = 2)  # Example LD matrix
beta <- c(1, 2)                                 # Example genotype effects
RL <- get_lower_chol(R)                         # Compute lower Cholesky decomposition
compute_s2g(RL, beta)                            # Compute genetic variance
}
