% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_transQTL_null.R
\name{gene_data_null1}
\alias{gene_data_null1}
\title{Generate null data by random shuffling of genotypes}
\usage{
gene_data_null1(data_h1)
}
\arguments{
\item{data_h1}{A dataset generated by the gene_data function}
}
\value{
A modified dataset with randomly shuffled genotypes
}
\description{
This function creates a null dataset by randomly shuffling the genotype data
while maintaining the original gene expression data.
}
\examples{
# Generate original data
data_h1 <- gene_data(n = 1000, p = 60, A_snp_cis, A_cis_trans, A_trans)

# Generate null data by shuffling genotypes
data_null <- gene_data_null1(data_h1)

}
