% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_transQTL_null.R
\name{gene_data_null2}
\alias{gene_data_null2}
\title{Generate null data by removing cis-to-trans effects}
\usage{
gene_data_null2(data_h1, A_trans, gamma_trans_trans = 0.5)
}
\arguments{
\item{data_h1}{A dataset generated by the gene_data function}

\item{A_trans}{Binary adjacency matrix indicating trans-trans gene relationships}

\item{gamma_trans_trans}{Effect size for trans to trans gene effects}
}
\value{
A modified dataset with new trans gene expression lacking cis-to-trans effects
}
\description{
This function creates a null dataset by regenerating trans gene expression data
without the influence of cis genes, while maintaining trans-to-trans effects.
}
\examples{
# Generate original data
data_h1 <- gene_data(n = 1000, p = 60, A_snp_cis, A_cis_trans, A_trans)

# Generate null data without cis-to-trans effects
data_null <- gene_data_null2(data_h1, A_trans, gamma_trans_trans = 0.5)

}
