% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_transQTL_null.R
\name{gene_data_null3}
\alias{gene_data_null3}
\title{Generate null data using ARCHIE setting}
\usage{
gene_data_null3(n, p, data_h1, A_snp_cis, A_cis_trans, A_trans)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of SNPs}

\item{data_h1}{A dataset generated by the gene_data function}

\item{A_snp_cis}{Binary adjacency matrix indicating SNP-cis gene relationships}

\item{A_cis_trans}{Binary adjacency matrix indicating cis-trans gene relationships}

\item{A_trans}{Binary adjacency matrix indicating trans-trans gene relationships}
}
\value{
A matrix of Z-statistics under the null hypothesis
}
\description{
This function creates a null dataset by directly generating a Z-statistic matrix
under the multivariate normal assumption used in the ARCHIE method.
}
\examples{
# Generate original data
data_h1 <- gene_data(n = 1000, p = 60, A_snp_cis, A_cis_trans, A_trans)

# Generate null Z-matrix using ARCHIE setting
Z_matrix_null <- gene_data_null3(1000, 60, data_h1, A_snp_cis, A_cis_trans, A_trans)

}
