% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linreg.R
\name{get_ld}
\alias{get_ld}
\title{Compute LD Matrix}
\usage{
get_ld(G, ld_ridge = 0.1)
}
\arguments{
\item{G}{Genotype matrix.}

\item{ld_ridge}{Ridge regularization parameter for LD matrix.}
}
\value{
LD matrix.
}
\description{
Computes the LD matrix from genotype data, adjusting for minor allele frequencies,
regularizing to ensure positive semi-definiteness, and standardizing to a correlation matrix.
}
\examples{
# Random genotype matrix as an example
G <- matrix(rbinom(1000, 2, 0.5), nrow = 100, ncol = 10)
ld_ridge <- 0.1
LD <- get_ld(G, ld_ridge)
}
