% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linreg.R
\name{get_lower_chol}
\alias{get_lower_chol}
\title{Compute Lower Cholesky Decomposition}
\usage{
get_lower_chol(R)
}
\arguments{
\item{R}{The LD matrix for which the lower Cholesky decomposition is to be computed.}
}
\value{
Lower Cholesky factor of the LD matrix.
}
\description{
Computes the lower Cholesky decomposition of the LD matrix.
}
\examples{
R <- matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2)  # Example LD matrix
L <- get_lower_chol(R)                              # Compute lower Cholesky decomposition
}
