% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_transQTL.R
\name{get_random_A}
\alias{get_random_A}
\title{Generate random adjacency matrices for trans gene associations}
\usage{
get_random_A(A_cell_1, p = c(1/3, 2/3))
}
\arguments{
\item{A_cell_1}{List containing fixed adjacency matrices for cell-type 1 (A_cis_trans and A_trans).}

\item{p}{Vector of probabilities for the associations.}
}
\value{
A list containing random adjacency matrices for cell-type 2 (A_cis_trans and A_trans).
}
\description{
This function creates random adjacency matrices for cis-trans and trans-trans effects
based on specified probabilities, as illustrated in Figure 7(B) and (C).
}
\examples{
A_cell_1 <- list(A_cis_trans = matrix(1, nrow = 8, ncol = 9),
                 A_trans = matrix(0, nrow = 9, ncol = 9))
p <- c(1/4, 1/2, 1/2) # Example probabilities
A_cell_2 <- get_random_A(A_cell_1, p)
}
