% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linreg.R
\name{sim_beta}
\alias{sim_beta}
\title{Simulate QTL Effects}
\usage{
sim_beta(RL, ncausal, eqtl_h2, rescale = TRUE)
}
\arguments{
\item{RL}{Lower Cholesky factor of the p x p LD matrix for the population.}

\item{ncausal}{Object of class NumCausalSNPs containing the number of causal SNPs to select.}

\item{eqtl_h2}{The heritability of gene expression.}

\item{rescale}{Logical indicating whether to rescale effects such that var(b V b) = h2 (default is TRUE).}
}
\value{
Vector of causal effects.
}
\description{
Sample QTL effects under a specified architecture.
}
\examples{
# Create a correlation matrix as R
R <- matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2)
ncausal <- list(value = 50, is_pct = TRUE)  # 50\% of observed SNPs
eqtl_h2 <- 0.5                              # Heritability of gene expression
RL <- get_lower_chol(R)                         # Compute lower Cholesky decomposition
sim_beta(RL, ncausal, eqtl_h2)               # Simulate QTL effects
}
