% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linreg.R
\name{sim_beta_fix_variant}
\alias{sim_beta_fix_variant}
\title{Simulate Multiple Traits from Genotype and Assigned Causal Index Vector}
\usage{
sim_beta_fix_variant(G, causal_index, ntrait = 1, is_h2g_total = TRUE)
}
\arguments{
\item{G}{Genotype matrix.}

\item{causal_index}{Vector of indices for causal variants.}

\item{ntrait}{Number of traits to simulate.}

\item{is_h2g_total}{Logical indicating if h2g is total (TRUE) or per-SNP (FALSE).}
}
\value{
Matrix of effect sizes (variants × traits).
}
\description{
Simulate Multiple Traits from Genotype and Assigned Causal Index Vector
}
