% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linreg.R
\name{sim_geno}
\alias{sim_geno}
\title{Simulate Genotypes}
\usage{
sim_geno(RL, n)
}
\arguments{
\item{RL}{Lower Cholesky factor of the p x p LD matrix for the population.}

\item{n}{Number of genotypes to sample.}
}
\value{
Centered and scaled genotype matrix.
}
\description{
Sample genotypes from a multivariate normal approximation.
}
\examples{
R <- matrix(c(1, 0.5, 0.5, 1), nrow = 2)  # LD matrix
n <- 100                                  # Number of genotypes to sample
RL <- get_lower_chol(R)                              # Compute lower Cholesky decomposition
G <- sim_geno(RL, n)                       # Simulate genotypes
}
