% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_genotype.R
\name{sim_geno_indep}
\alias{sim_geno_indep}
\title{Simulate Independent Genotypes}
\usage{
sim_geno_indep(n, p, min_maf = 0.01, max_maf = 0.4, scale = FALSE)
}
\arguments{
\item{n}{Sample size.}

\item{p}{Number of SNPs.}

\item{min_maf}{Minimum minor allele frequency.}

\item{max_maf}{Maximum minor allele frequency.}

\item{scale}{Logical, whether to scale the data.}
}
\value{
A matrix of genotypes.
}
\description{
Simulate Independent Genotypes
}
\examples{
sim_geno_indep(n = 100, p = 10, min_maf = 0.01, max_maf = 0.4, scale = TRUE)
}
