% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_genotype.R
\name{sim_geno_LD}
\alias{sim_geno_LD}
\title{Simulate Genotypes Based on LD Matrix}
\usage{
sim_geno_LD(
  n,
  LD,
  min_maf = 0.01,
  max_maf = 0.4,
  lambda = 0.001,
  is.discrete = FALSE,
  scale = FALSE,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{n}{Sample size.}

\item{LD}{LD matrix.}

\item{min_maf}{Minimum minor allele frequency.}

\item{max_maf}{Maximum minor allele frequency.}

\item{lambda}{Regularization parameter.}

\item{is.discrete}{Logical, whether to generate discrete (TRUE) or continuous (FALSE) genotypes.}

\item{scale}{Logical, whether to scale the data using scale_faster.}

\item{tol}{Tolerance for checking diagonal elements of LD matrix.}
}
\value{
A matrix of genotypes.
}
\description{
Simulate Genotypes Based on LD Matrix
}
\examples{
sim_geno_LD(n = 100, LD = matrix(runif(100), 10, 10), min_maf = 0.01, max_maf = 0.4, lambda = 1e-3, is.discrete = FALSE, scale = TRUE)
}
