% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_genotype.R
\name{sim_geno_real}
\alias{sim_geno_real}
\title{Simulate Genotypes Based on Real Data}
\usage{
sim_geno_real(n, file_path, min_maf = 0.01, scale = FALSE)
}
\arguments{
\item{n}{Sample size.}

\item{file_path}{Path to the genotype file.}

\item{min_maf}{Minimum minor allele frequency.}

\item{scale}{Logical, whether to scale the data.}
}
\value{
A matrix of genotypes.
}
\description{
Simulate Genotypes Based on Real Data
}
\examples{
sim_geno_real(n = 100, file_path = "path/to/real/file", min_maf = 0.01, scale = TRUE)
}
