% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_genotype.R
\name{sim_geno_UKB}
\alias{sim_geno_UKB}
\title{Simulate Genotypes Based on UK Biobank}
\usage{
sim_geno_UKB(n, file_path, min_maf = 0.01, scale = FALSE)
}
\arguments{
\item{n}{Sample size.}

\item{file_path}{Path to the UK Biobank file.}

\item{min_maf}{Minimum minor allele frequency.}

\item{scale}{Logical, whether to scale the data using scale_faster.}
}
\value{
A matrix of genotypes.
}
\description{
Simulate Genotypes Based on UK Biobank
}
\examples{
sim_geno_UKB(n = 100, file_path = "path/to/UKB/file", min_maf = 0.01, scale = TRUE)
}
