% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linreg.R
\name{sim_multi_traits}
\alias{sim_multi_traits}
\title{Simulate Multiple Traits from Genotype and Effect Sizes}
\usage{
sim_multi_traits(
  G,
  B,
  h2g,
  is_h2g_total = TRUE,
  max_h2g = 0.8,
  residual_corr = NULL,
  scale_Y = FALSE
)
}
\arguments{
\item{G}{Genotype matrix.}

\item{B}{Matrix of effect sizes for multiple traits.}

\item{h2g}{Heritability (proportion of variance explained by genetics).}

\item{is_h2g_total}{Logical indicating if h2g is total (TRUE) or per-SNP (FALSE).}

\item{max_h2g}{Maximum heritability allowed in the model (default is 0.8).}

\item{residual_corr}{Matrix of residual correlations (NULL for independent samples).}

\item{scale_Y}{Logical indicating if the phenotypes should be scaled (default is FALSE).}
}
\value{
A list containing the simulated phenotypes (`Y`) and residual variance (`residual_var`).
}
\description{
This function simulates multiple traits (phenotypes) based on genotype data, 
a matrix of effect sizes, and heritability. It allows specifying if the heritability
is total or per-SNP, optionally scales the phenotypes, and can handle residual correlations.
per eQTL heritability is 0.05 to 0.07 according to https://www.nature.com/articles/ng.3506
}
\examples{
G <- matrix(rnorm(1000), nrow = 100, ncol = 10) # Random genotype matrix
# Simulating effect sizes for two traits
B <- sapply(1:2, function(i) sim_beta(RL, ncausal, eqtl_h2)) 
h2g <- 0.5                                      # Heritability
result <- sim_multi_traits(G, B, h2g)           # Simulate multiple traits
}
