% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_eQTL.R
\name{simulate_infinitesimal_effects}
\alias{simulate_infinitesimal_effects}
\title{Simulate Infinitesimal Effects for eQTL Data}
\usage{
simulate_infinitesimal_effects(G, h2_infinitesimal, infinitesimal_indices)
}
\arguments{
\item{G}{A standardized genotype matrix.}

\item{h2_infinitesimal}{Heritability allocated to infinitesimal effects.}

\item{infinitesimal_indices}{Indices of SNPs to receive infinitesimal effects.}
}
\value{
A vector of effect sizes (with zeros outside infinitesimal_indices).
}
\description{
This function simulates the infinitesimal (polygenic) background effects for
a set of SNPs.
}
\keyword{internal}
