% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linreg.R
\name{simulate_polygenic_trait}
\alias{simulate_polygenic_trait}
\title{Simulate a Polygenic Trait}
\usage{
simulate_polygenic_trait(g, h2g)
}
\arguments{
\item{g}{Vector of latent genetic values.}

\item{h2g}{Heritability of the trait in the population.}
}
\value{
Simulated phenotype.
}
\description{
Simulate a complex trait as a function of latent genetic values and environmental noise.
}
\examples{
R <- matrix(c(1, 0.5, 0.5, 1), nrow = 2)  # LD matrix
n <- 100                                  # Number of samples
G <- sim_geno_LD(n, R)                     # Simulate genotypes
ncausal <- 1                               # Number of causal SNPs
b <- sim_beta(G, ncausal, ntrait = 1, is_h2g_total = TRUE, shared_pattern = "all")
g <- G \%*\% b                               # latent genetic values
y <- simulate_polygenic_trait(g, h2g = 0.5)  # Simulate the complex trait
}
