% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_transQTL.R
\name{simulate_trans_mixture_celltype}
\alias{simulate_trans_mixture_celltype}
\title{Simulate trans gene expression for a mixture of cell types}
\usage{
simulate_trans_mixture_celltype(
  E_cis,
  phi_gene,
  A_cell_1,
  A_cell_2 = NULL,
  omega = NULL,
  p = c(1/3, 2/3)
)
}
\arguments{
\item{E_cis}{Matrix of cis gene expressions with dimensions n x m.}

\item{phi_gene}{Per-gene heritability.}

\item{A_cell_1}{List containing adjacency matrices for cell-type 1 (A_cis_trans and A_trans).}

\item{A_cell_2}{Optionally, a similar list for cell-type 2.}

\item{omega}{Optionally, a weight for combining cell types. If NULL, drawn from U(0.2, 0.8).}
}
\value{
A list containing matrices E_trans, E_trans_cell_1, and E_trans_cell_2.
}
\description{
This function simulates trans gene expression considering different cell types,
each with its own gene-gene regulatory network.
}
\examples{
n <- 1000
m <- 8
g <- 9
phi_gene <- 0.20
E_cis <- matrix(rnorm(n * m), n, m)
A_cell_1 <- list(A_cis_trans = matrix(sample(0:1, m * g, replace = TRUE), nrow = m),
                A_trans = matrix(sample(0:1, g * g, replace = TRUE), nrow = g))
result <- simulate_trans_mixture_celltype(E_cis, phi_gene, A_cell_1)
}
