% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linreg.R
\name{sim_beta}
\alias{sim_beta}
\title{Simulate QTL Effects}
\usage{
sim_beta(G, ncausal, ntrait = 1, is_h2g_total = TRUE, shared_pattern = "all")
}
\arguments{
\item{G}{Genotype matrix}

\item{ncausal}{Output from function parse_num_causal_snps, how many variants have non-negative effects (being causal)}

\item{ntrait}{Number of simulated phenotypes (traits)}

\item{is_h2g_total}{Logical indicating if h2g is total (TRUE) or per-SNP (FALSE).}

\item{shared_pattern}{if is "all", all traits will have the same causal variant(s) with non-zero effect. if is "random", all traits will have independent (random) causal variant(s)}
}
\value{
Matrix of causal effect sizes (variants × traits).
}
\description{
Sample QTL effects under a specified architecture.
}
\examples{
# Create a genotype matrix
G = matrix(rbinom(1000, 2, 0.5), nrow = 1000, ncol = 50) 
B = sim_beta(G, ncausal = 5, ntrait = 3, is_h2g_total = T, shared_pattern = "all")
B = sim_beta(G, ncausal = 1, ntrait = 5, is_h2g_total = F, shared_pattern = "random")
}
