% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_linreg.R
\name{simulate_polygenic_trait}
\alias{simulate_polygenic_trait}
\title{Simulate a Polygenic Trait}
\usage{
simulate_polygenic_trait(g, h2g)
}
\arguments{
\item{g}{Vector of latent genetic values.}

\item{h2g}{Heritability of the trait in the population.}
}
\value{
Simulated phenotype.
}
\description{
Simulate a complex trait as a function of latent genetic values and environmental noise.
}
\examples{
R <- matrix(c(1, 0.5, 0.5, 1), nrow = 2)  # LD matrix
n <- 100                                  # Number of genotypes to sample
RL <- get_lower_chol(R)                              # Compute lower Cholesky decomposition
G <- sim_geno(RL, n)                       # Simulate genotypes
ncausal <- list(value = 50, is_pct = TRUE)  # 50\% of observed SNPs
eqtl_h2 <- 0.5                              # Heritability of gene expression
b <- sim_beta(RL, ncausal, eqtl_h2) 
g <- G \%*\% b                   # latent genetic values
y <- simulate_polygenic_trait(g, eqtl_h2)  # Simulate the complex trait
}
